Explode Upload class into the places which actually need the code within.
This also changes the behavior of the ChipCsvImport path, to be less magical
